requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.2
python-dotenv==1.0.1
spotipy==2.24.0
pydantic-settings==2.4.0
